backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
python-socketio>=5.11
python-dotenv>=1.0
pydantic>=2.0
mistralai>=1.0
